refactor: move all deal with fees logic into common crate
